+++ sdk/src/main/java/ly/count/android/sdk/ModuleRemoteConfig.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import ly.count.android.sdk.internal.RemoteConfigHelper;
 import ly.count.android.sdk.internal.RemoteConfigValueStore;
 import org.json.JSONException;
@@ -734,16 +735,78 @@ public class ModuleRemoteConfig extends ModuleBase {
         }
 
         /**
+         * Returns all available remote config values and enrolls to A/B tests for those values
+         *
+         * @return The available RC values
+         */
+        public @NonNull Map<String, RCData> getAllValuesAndEnroll() {
+            synchronized (_cly) {
+                L.i("[RemoteConfig] Getting all Remote config values v2 and enrolling");
+                Map<String, RCData> values = getAllRemoteConfigValuesInternal();
+
+                if (!values.isEmpty()) {
+
+                    Set<String> setOfKeys = values.keySet();
+                    String[] arrayOfKeys = new String[setOfKeys.size()];
+
+                    // set to array
+                    int i = 0;
+                    for (String key : setOfKeys)
+                        arrayOfKeys[i++] = key;
+
+                    // enroll
+                    enrollIntoABTestsForKeys(arrayOfKeys);
+                } else {
+                    L.w("[RemoteConfig] No value to enroll");
+                }
+
+                return values;
+            }
+        }
+
+        /**
          * Return the remote config value for a specific key
          *
          * @param key Key for which the remote config value needs to be returned
          * @return The returned value. If no value existed for the key then the inner object will be returned as "null"
          */
-        public @NonNull RCData getValue(@Nullable String key) {
+        public @NonNull RCData getValue(@NonNull String key) {
             synchronized (_cly) {
+                if (key.equals("")) {
+                    L.w("[RemoteConfig] A valid key should be provided to get its value.");
+                    return new RCData(null, true);
+                }
+
                 L.i("[RemoteConfig] Getting Remote config values for key:[" + key + "] v2");
 
                 return getRCValue(key);
+            }
+        }
+
+        /**
+         * Returns the remote config value for a specific key and enrolls to A/B tests for it
+         *
+         * @param key Key for which the remote config value needs to be returned
+         * @return The returned value. If no value existed for the key then the inner object will be returned as "null"
+         */
+        public @NonNull RCData getValueAndEnroll(@NonNull String key) {
+            synchronized (_cly) {
+                if (key.equals("")) {
+                    L.w("[RemoteConfig] A valid key should be provided to get its value.");
+                    return new RCData(null, true);
+                }
+
+                L.i("[RemoteConfig] Getting Remote config values for key:[" + key + "] v2 and enrolling");
+
+                RCData value = getRCValue(key);
+
+                if (value.value != null) {
+                    String[] arrayOfKeys = new String[] { key };
+                    enrollIntoABTestsForKeys(arrayOfKeys);
+                } else {
+                    L.w("[RemoteConfig] No value to enroll");
+                }
+                return value;
             }
         }